Day 23 part 2
Final runtime: ~0.5s

That took a while to figure out!

Naive solution with array copying would take ~4 hours,
acceptable but I felt we could do better!

Second version used linked list which was even slower,
likely because of the average O(n) lookup time which
is needed every step to find the destination

First idea was to speed that up with a lookup table,
but since the key values are just the labels 1 -> 1000000
this can simply be an array. So in the end the data structure
ended up quite simple and elegant! Love it!